#5 refactor : refactor dto and change file location
