+++ src/main/java/com/miniproject/backend/loanproduct/dto/LoanRateResponseDTO.java
@@ -1,4 +1,4 @@
-package com.miniproject.backend.shoppingbasket.dto;
+package com.miniproject.backend.loanproduct.dto;
 
 import com.miniproject.backend.loanproduct.domain.LoanRate;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -6,12 +6,12 @@ import lombok.Getter;
 
 @Getter
 @Schema(name = "대출금리 출력데이터")
-public class LoanRateDTO {
+public class LoanRateResponseDTO {
 
     @Schema(name = "대출금리")
     private float avgRate;
 
-    public LoanRateDTO(LoanRate loanRate){
+    public LoanRateResponseDTO(LoanRate loanRate){
         this.avgRate = loanRate.getAvgRate();
     }
